requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.0
dj-database-url==0.5.0
Django==4.0.3
django-bootstrap-modal-forms==2.2.0
django-extra-views==0.13.0
django-funky-sheets==0.2.0
django-heroku==0.3.1
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.0.4
Pillow==9.0.1
psycopg2==2.8.5
pytz==2019.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
Unipath==1.1
whitenoise==5.0.1
